Refactor: config 패키지를 community 패키지 하위로 옮김
