+++ src/main/java/com/filmdoms/community/account/controller/AccountController.java
@@ -1,6 +1,6 @@
 package com.filmdoms.community.account.controller;
 
-import com.filmdoms.community.account.config.jwt.JwtTokenProvider;
+import com.filmdoms.community.config.jwt.JwtTokenProvider;
 import com.filmdoms.community.account.data.constant.AccountRole;
 import com.filmdoms.community.account.data.dto.AccountDto;
 import com.filmdoms.community.account.data.dto.LoginDto;
