+++ src/main/java/com/filmdoms/community/config/PasswordEncoderConfig.java
@@ -1,4 +1,4 @@
-package com.filmdoms.community.account.config;
+package com.filmdoms.community.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
