+++ src/main/java/com/filmdoms/community/config/oauth/CustomOAuthFailureHandler.java
@@ -1,11 +1,10 @@
-package com.filmdoms.community.account.config.oauth;
+package com.filmdoms.community.config.oauth;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.filmdoms.community.account.config.jwt.JwtTokenProvider;
+import com.filmdoms.community.config.jwt.JwtTokenProvider;
 import com.filmdoms.community.account.data.dto.response.Response;
 import com.filmdoms.community.account.exception.ApplicationException;
 import com.filmdoms.community.account.exception.ErrorCode;
-import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
