+++ src/test/java/com/filmdoms/community/banner/controller/BannerControllerTest.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.filmdoms.community.banner.data.dto.request.BannerRequestDto;
 import com.filmdoms.community.banner.data.dto.response.BannerResponseDto;
 import com.filmdoms.community.banner.service.BannerService;
-import com.filmdoms.community.config.TestSecurityConfig;
-import com.filmdoms.community.config.annotation.WithMockAdminUser;
+import com.filmdoms.community.testconfig.TestSecurityConfig;
+import com.filmdoms.community.testconfig.annotation.WithMockAdminUser;
 import com.filmdoms.community.file.data.dto.response.FileResponseDto;
 import com.filmdoms.community.file.data.entity.File;
 import org.junit.jupiter.api.DisplayName;
