+++ src/test/java/com/filmdoms/community/testconfig/TestSecurityConfig.java
@@ -1,10 +1,10 @@
-package com.filmdoms.community.config;
+package com.filmdoms.community.testconfig;
 
-import com.filmdoms.community.account.config.SecurityConfig;
-import com.filmdoms.community.account.config.jwt.JwtTokenProvider;
-import com.filmdoms.community.account.config.oauth.CustomOAuth2AuthorizationRequestResolver;
-import com.filmdoms.community.account.config.oauth.CustomOAuthFailureHandler;
-import com.filmdoms.community.account.config.oauth.CustomOAuthSuccessHandler;
+import com.filmdoms.community.config.SecurityConfig;
+import com.filmdoms.community.config.jwt.JwtTokenProvider;
+import com.filmdoms.community.config.oauth.CustomOAuth2AuthorizationRequestResolver;
+import com.filmdoms.community.config.oauth.CustomOAuthFailureHandler;
+import com.filmdoms.community.config.oauth.CustomOAuthSuccessHandler;
 import com.filmdoms.community.account.data.constant.AccountRole;
 import com.filmdoms.community.account.data.entity.Account;
 import com.filmdoms.community.account.repository.AccountRepository;
